% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{llkhd}
\alias{llkhd}
\title{Calculate constrained likelihood}
\usage{
llkhd(x, qn, nrisk, nevent, event_time, event_ind)
}
\arguments{
\item{x}{Time we are interested in comparing.}

\item{qn}{Parameter we are optimizing which survival probability at time x is exp(qn)}

\item{nrisk}{A vector of number at risk for each event time}

\item{nevent}{A vector representing the number of event}

\item{event_time}{A vector of event times.}

\item{event_ind}{A vector of event indicator (1, 0)}
}
\value{
Likelihood value
}
\description{
Calculate constrained likelihood for further obtaining the constrained NPMLE
}
\details{
DETAILS
}
\examples{
library(constrselect)
llkhd(x=6,qn = -0.1,nrisk = c(15, 14, 13, 12, 11, 10),nevent = c(1, 1, 1, 1, 1, 1),
event_time = c(0.47, 0.84, 2.25, 2.5, 2.8, 4.5, 6, 6.16, 6.6, 6.86, 8.17),
event_ind = c(1, 1, 1, 1, 1, 1, 0, 0,0,0,0))  ## -24.12788
}
