% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Mg}
\alias{Mg}
\title{FUNCTION_TITLE}
\usage{
Mg(x, event_time, event_ind)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{event_time}{PARAM_DESCRIPTION}

\item{event_ind}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
