% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Ng}
\alias{Ng}
\title{Calculate number at risk at a specific time}
\usage{
Ng(x, event_time)
}
\arguments{
\item{x}{Time we are interested in comparing.}

\item{event_time}{A vector of event times.}
}
\value{
A number at risk at time x
}
\description{
Calculate number at risk at time x, which will be used in the following likelihood calculation
}
\details{
DETAILS
}
\examples{
library(constrselect)
Ng(x=6,event_time=c(0.41, 0.8, 2.2, 2.5, 3, 4.5, 6, 6.2, 6.5, 7)) ## 4
}
