% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{order_constrain}
\alias{order_constrain}
\title{FUNCTION_TITLE}
\usage{
order_constrain(r, n, order_list)
}
\arguments{
\item{r}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION}

\item{order_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[quadprog]{solve.QP}}
}
