% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pickwin_bin_exact}
\alias{pickwin_bin_exact}
\title{FUNCTION_TITLE}
\usage{
pickwin_bin_exact(
  n,
  p1,
  strata_diff,
  D = c(0.15, 0.15),
  d = c(0.05, 0.05),
  prop.strat = 0.3,
  study = NULL
)
}
\arguments{
\item{n}{PARAM_DESCRIPTION}

\item{p1}{PARAM_DESCRIPTION}

\item{strata_diff}{PARAM_DESCRIPTION}

\item{D}{PARAM_DESCRIPTION, Default: c(0.15, 0.15)}

\item{d}{PARAM_DESCRIPTION, Default: c(0.05, 0.05)}

\item{prop.strat}{PARAM_DESCRIPTION, Default: 0.3}

\item{study}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
