% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pickwin_bin_multiple}
\alias{pickwin_bin_multiple}
\title{FUNCTION_TITLE}
\usage{
pickwin_bin_multiple(
  n,
  pa_list,
  D = c(0.15, 0.15, 0.15),
  d_diff = c(0.05, 0.05, 0.05),
  prop.strat = c(0.2, 0.3, 0.5),
  study = "Constrained",
  S,
  cluster = 6,
  order_list,
  with_seed = NULL
)
}
\arguments{
\item{n}{PARAM_DESCRIPTION}

\item{pa_list}{PARAM_DESCRIPTION}

\item{D}{PARAM_DESCRIPTION, Default: c(0.15, 0.15, 0.15)}

\item{d_diff}{PARAM_DESCRIPTION, Default: c(0.05, 0.05, 0.05)}

\item{prop.strat}{PARAM_DESCRIPTION, Default: c(0.2, 0.3, 0.5)}

\item{study}{PARAM_DESCRIPTION, Default: 'Constrained'}

\item{S}{PARAM_DESCRIPTION}

\item{cluster}{PARAM_DESCRIPTION, Default: 6}

\item{order_list}{PARAM_DESCRIPTION}

\item{with_seed}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
