% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pickwin_surv_fun}
\alias{pickwin_surv_fun}
\title{FUNCTION_TITLE}
\usage{
pickwin_surv_fun(
  maxn,
  prop,
  event_rate_A,
  trt_diff,
  d_diff,
  arrival_rate,
  FUP,
  x,
  S,
  study = "Constrained",
  cluster,
  order_list,
  with_seed = NULL
)
}
\arguments{
\item{maxn}{PARAM_DESCRIPTION}

\item{prop}{PARAM_DESCRIPTION}

\item{event_rate_A}{PARAM_DESCRIPTION}

\item{trt_diff}{PARAM_DESCRIPTION}

\item{d_diff}{PARAM_DESCRIPTION}

\item{arrival_rate}{PARAM_DESCRIPTION}

\item{FUP}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{S}{PARAM_DESCRIPTION}

\item{study}{PARAM_DESCRIPTION, Default: 'Constrained'}

\item{cluster}{PARAM_DESCRIPTION}

\item{order_list}{PARAM_DESCRIPTION}

\item{with_seed}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[doParallel]{registerDoParallel}}
\code{\link[foreach]{foreach}}
\code{\link[survival]{survfit}}
}
