% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_surv}
\alias{sim_surv}
\title{FUNCTION_TITLE}
\usage{
sim_surv(nmax, arrival_rate, event_rate, FUP)
}
\arguments{
\item{nmax}{PARAM_DESCRIPTION}

\item{arrival_rate}{PARAM_DESCRIPTION}

\item{event_rate}{PARAM_DESCRIPTION}

\item{FUP}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
