% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{llkhd}
\alias{llkhd}
\title{FUNCTION_TITLE}
\usage{
llkhd(qq, nrisk, nevent, event_time, event_ind)
}
\arguments{
\item{qq}{PARAM_DESCRIPTION}

\item{nrisk}{PARAM_DESCRIPTION}

\item{nevent}{PARAM_DESCRIPTION}

\item{event_time}{PARAM_DESCRIPTION}

\item{event_ind}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
