% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Ng}
\alias{Ng}
\title{FUNCTION_TITLE}
\usage{
Ng(x, event_time)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{event_time}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
