% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{partial_order}
\alias{partial_order}
\title{FUNCTION_TITLE}
\usage{
partial_order(order_list)
}
\arguments{
\item{order_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
