% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pickwin_surv_fun}
\alias{pickwin_surv_fun}
\title{Estimate constrained NPMLE of survival probabilities in stratified selection setting}
\usage{
pickwin_surv_fun(
  maxn,
  prop,
  event_rate_A,
  trt_diff,
  d,
  arrival_rate,
  FUP,
  x,
  S,
  study = "Constrained",
  cluster,
  order_list,
  with_seed = NULL
)
}
\arguments{
\item{maxn}{Number of patients need to be accrued for each treatment arm}

\item{prop}{The sample size proportion for two strata, Default: c(0.2, 0.3, 0.5)}

\item{event_rate_A}{The exponential event rate for patients in the inferior treatment arm}

\item{trt_diff}{A vector of two treatment arms differences for each stratum, Default: c(0.1, 0.1,0.1)}

\item{d}{A vector of ambiguous region for each stratum, Default: c(0.05, 0.05, 0.05)}

\item{arrival_rate}{The poisson arrival rate for patients, number of patients accrued each month/year}

\item{FUP}{Additional follow up time after the last patient is accrued}

\item{x}{Time we are interested in comparing.}

\item{S}{Number of simulation}

\item{study}{Could be either "Constrained" or "Origin" for the two type of study design, Default: 'Constrained'}

\item{cluster}{Number of parallel running CPU cores, Default: 6}

\item{order_list}{A list of strata order allowing for partial ordering grouped in a vector within a list.}

\item{with_seed}{Random seed for simulation, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Estimate constrained NPMLE of survival probabilities in stratified selection setting using simulation.
For each simulation setting, get whether it is correctly or wrongly counted in.
}
\details{
DETAILS
}
\examples{
library(constrselect)
test <- pickwin_surv_fun(maxn=50,prop=c(0.3,0.3,0.4),event_rate_A=c(0.08,0.05, 0.05),
trt_diff=c(0.1,0.1,0.1),d_diff=c(0.05,0.05,0.05), arrival_rate=4,FUP=6,
x=6,S=100,study = "Constrained",cluster=6,order_list=list(1,c(2,3)),with_seed = 111)
}
\seealso{
\code{\link[doParallel]{registerDoParallel}}
\code{\link[foreach]{foreach}}
\code{\link[survival]{survfit}}
}
