% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_surv}
\alias{sim_surv}
\title{Simulation of survival times}
\usage{
sim_surv(nmax, arrival_rate, event_rate, FUP)
}
\arguments{
\item{nmax}{Number of patients need to be accrued.}

\item{arrival_rate}{The Poisson arrival rate for patients, number of patients accrued each month/year.}

\item{event_rate}{The exponential event rate for patients.}

\item{FUP}{Additional follow up time after the last patient is accrued.}
}
\value{
Return a dataframe with time to event/censoring and event status (1,0).
}
\description{
Simulate survival times based on Poisson arrival and exponential survival.
Follow up the patients for addition month after the last patient is accrued.
}
\details{
DETAILS
}
\examples{
library(constrselect)
sim_surv(nmax=12,arrival_rate=4,event_rate=0.08,FUP=6)
}
